move it to ConfigUtils
